refactor: extract createUser method to UserService
