+++ src/main/java/com/fdmgroup/forex/services/AuthUserService.java
@@ -3,7 +3,6 @@ package com.fdmgroup.forex.services;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.fdmgroup.forex.models.User;
@@ -17,11 +16,9 @@ import com.fdmgroup.forex.security.AuthUser;
 public class AuthUserService implements UserDetailsService {
 
 	private UserRepo userRepo;
-	private PasswordEncoder pwdEncoder;
 
-	public AuthUserService(UserRepo userRepo, PasswordEncoder pwdEncoder) {
+	public AuthUserService(UserRepo userRepo) {
 		this.userRepo = userRepo;
-		this.pwdEncoder = pwdEncoder;
 	}
 
 	@Override
@@ -30,11 +27,6 @@ public class AuthUserService implements UserDetailsService {
 		User user = this.userRepo.findByUsername(username).orElseThrow(
 				() -> new UsernameNotFoundException(username));
 		return new AuthUser(user);
-	}
-
-	public void createUser(User user) {
-		user.setPassword(pwdEncoder.encode(user.getPassword()));
-		userRepo.save(user);
 	}
 
 }
